requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
chumpy==0.70
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
cmake==3.25.0
colorlog==6.9.0
cycler==0.12.1
decorator==5.2.1
einops==0.6.1
exceptiongroup==1.3.0
executing==2.2.0
ffmpy==0.3.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
ftfy==6.1.1
gdown==4.7.1
grpcio==1.54.2
h11==0.14.0
h5py==3.14.0
hf-xet==1.1.5
huggingface-hub==0.33.1
hydra-core==1.3.2
idna==3.10
importlib-metadata==5.0.0
importlib-resources==5.12.0
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
lightning-utilities==0.14.3
lit==15.0.7
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.1.3
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.6.3
natsort==8.4.0
networkx==3.3
numpy==1.21.5
omegaconf==2.3.0
packaging==25.0
pandas==2.0.3
parso==0.8.4
pexpect==4.9.0
Pillow==9.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.10.1
six==1.17.0
smplx==0.1.28
sniffio==1.3.0
soupsieve==2.7
stack-data==0.6.3
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.2
tomli==2.2.1
torch==2.1.0+cu118
torch-tb-profiler==0.4.3
torchaudio==2.1.0+cu118
torchmetrics==1.7.4
torchvision==0.16.0+cu118
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.4
trimesh==4.6.11
triton==2.1.0
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
vector-quantize-pytorch==1.6.30
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
